Removed PaperLib and make use of paper's native adventure methods. Only include and relocate minimessage because it was not native available on 1.17.
