+++ hyperverse-core/build.gradle.kts
@@ -26,7 +26,6 @@ dependencies {
     // TODO: Remove, because yuck.
     implementation("co.aikar:acf-paper:0.5.1-SNAPSHOT")
     implementation(libs.taskchain)
-    implementation(libs.paperlib)
     implementation(libs.guice) {
         exclude("com.google.guava", "guava")
     }
@@ -37,11 +36,9 @@ dependencies {
     implementation(libs.configurateHocon)
     implementation(libs.cloudPaper)
     implementation(libs.cloudMinecraftExtras)
-    implementation(libs.cloudMinecraftExtras)
 
     // TODO: Remove and use native versions.
-    implementation("net.kyori:adventure-platform-bukkit:4.3.2")
-    implementation("net.kyori:adventure-text-minimessage:4.14.0")
+    implementation("net.kyori:adventure-text-minimessage:4.18.0")
 
     implementation(projects.hyperverseNmsUnsupported)
     runtimeOnly(project(":hyperverse-nms-1-17")) {
@@ -140,14 +137,13 @@ tasks {
             }
         }
 
-        relocate("io.papermc.lib", "org.incendo.hyperverse.libs.paperlib")
         relocate("org.bstats", "org.incendo.hyperverse.libs.bstats")
         relocate("co.aikar.commands", "org.incendo.hyperverse.libs.aikar.commands")
         relocate("co.aikar.locales", "org.incendo.hyperverse.libs.aikar.locales")
         relocate("co.aikar.taskchain", "org.incendo.hyperverse.libs.taskchain")
         relocate("co.aikar.util", "org.incendo.hyperverse.libs.aikar.util")
         relocate("net.jodah.expiringmap", "org.incendo.hyperverse.libs.expiringmap")
-        relocate("net.kyori", "org.incendo.hyperverse.libs.kyori")
+        relocate("net.kyori.adventure.text.minimessage", "org.incendo.hyperverse.libs.minimessage")
         relocate("cloud.commandframework", "org.incendo.hyperverse.libs.cloud")
         relocate("org.spongepowered.configurate", "org.incendo.hyperverse.libs.configurate")
         relocate("io.leangen.geantyref", "org.incendo.hyperverse.libs.geantyref")
