+++ hyperverse-core/src/main/java/org/incendo/hyperverse/util/MessageUtil.java
@@ -17,13 +17,10 @@
 
 package org.incendo.hyperverse.util;
 
-import net.kyori.adventure.platform.bukkit.BukkitAudiences;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.plugin.java.JavaPlugin;
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.incendo.hyperverse.Hyperverse;
 import org.incendo.hyperverse.configuration.Message;
 import org.incendo.hyperverse.configuration.Messages;
 
@@ -34,7 +31,6 @@ import java.util.Objects;
  */
 public final class MessageUtil {
 
-    private static final BukkitAudiences AUDIENCES = BukkitAudiences.create(JavaPlugin.getPlugin(Hyperverse.class));
     private static final MiniMessage MINI_MESSAGE = MiniMessage.miniMessage();
 
     private MessageUtil() {
@@ -91,7 +87,7 @@ public final class MessageUtil {
             } else {
                 prefixedMessage = Messages.messagePrefixFancy.toString() + replacedMessage;
             }
-            AUDIENCES.sender(recipient).sendMessage(MINI_MESSAGE.deserialize(prefixedMessage));
+            recipient.sendMessage(MINI_MESSAGE.deserialize(prefixedMessage));
         } else {
             final String prefixedMessage = ChatColor.translateAlternateColorCodes(
                     '&',
