+++ hyperverse-nms-common/build.gradle.kts
@@ -4,7 +4,6 @@ plugins {
 
 dependencies {
     compileOnly(libs.paper)
-    compileOnlyApi(libs.paperlib)
     compileOnlyApi(libs.guice)
     compileOnlyApi(libs.taskchain)
 }
